+++ app/Http/Controllers/Backend/VendorProductVariantController.php
@@ -6,37 +6,28 @@ use App\DataTables\VendorProductVariantDataTable;
 use App\Http\Controllers\Controller;
 use App\Models\Product;
 use App\Models\ProductVariant;
+use App\Models\ProductVariantItem;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 
 class VendorProductVariantController extends Controller
 {
-    /**
-     * Display a listing of the resource.
-     */
+
     public function index(Request $request, VendorProductVariantDataTable $dataTable)
     {
         $product = Product::findOrFail($request->product);
-
         /** Check product vendor */
         if($product->vendor_id !== Auth::user()->id){
             abort(404);
         }
-
         return $dataTable->render('vendor.product.product-variant.index', compact('product'));
     }
 
-    /**
-     * Show the form for creating a new resource.
-     */
     public function create()
     {
         return view('vendor.product.product-variant.create');
     }
 
-    /**
-     * Store a newly created resource in storage.
-     */
     public function store(Request $request)
     {
         $request->validate([
@@ -50,22 +41,10 @@ class VendorProductVariantController extends Controller
         $varinat->name = $request->name;
         $varinat->status = $request->status;
         $varinat->save();
-
         toastr('Created Successfully!', 'success', 'success');
         return redirect()->route('vendor.products-variant.index', ['product' => $request->product]);
     }
 
-    /**
-     * Display the specified resource.
-     */
-    public function show(string $id)
-    {
-        //
-    }
-
-    /**
-     * Show the form for editing the specified resource.
-     */
     public function edit(string $id)
     {
         $variant = ProductVariant::findOrFail($id);
@@ -76,9 +55,6 @@ class VendorProductVariantController extends Controller
         return view('vendor.product.product-variant.edit', compact('variant'));
     }
 
-    /**
-     * Update the specified resource in storage.
-     */
     public function update(Request $request, string $id)
     {
         $request->validate([
@@ -94,27 +70,21 @@ class VendorProductVariantController extends Controller
         $varinat->name = $request->name;
         $varinat->status = $request->status;
         $varinat->save();
-
-        toastr('Updated Successfully!', 'success', 'success');
+        toastr('Updated Successfully!', 'info', 'success');
         return redirect()->route('vendor.products-variant.index', ['product' => $varinat->product_id]);
     }
 
-    /**
-     * Remove the specified resource from storage.
-     */
     public function destroy(string $id)
     {
         $varinat = ProductVariant::findOrFail($id);
-
         /** Check product vendor */
         if($varinat->product->vendor_id !== Auth::user()->id){
             abort(404);
         }
-
-        // $variantItemCheck = ProductVariantItem::where('product_variant_id', $varinat->id)->count();
-        // if($variantItemCheck > 0){
-        //     return response(['status' => 'error', 'message' => 'This variant contain variant items in it delete the variant items first for delete this variant!']);
-        // }
+        $variantItemCheck = ProductVariantItem::where('product_variant_id', $varinat->id)->count();
+        if($variantItemCheck > 0){
+            return response(['status' => 'error', 'message' => 'This variant contain variant items in it delete the variant items first for delete this variant!']);
+        }
         $varinat->delete();
         return response(['status' => 'success', 'message' => 'Deleted Successfully!']);
     }
